chore: tweak the layout of the application
